requirements.txt
async-generator==1.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
h11==0.14.0
idna==3.4
outcome==1.2.0
packaging==22.0
Pillow==9.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.28.1
selenium==4.5.0
Selenium-Screenshot==2.0.0
slack-sdk==3.19.1
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
webdriver-manager==3.8.5
wsproto==1.2.0
yapf==0.32.0
